# docs/OpenAPI.yaml
openapi: 3.0.3
info:
  title: DCS Server Management API
  version: 1.0.0
  description: >
    Standard REST API for managing DCS Dedicated Server hosting.
    Designed for an Orchestrator (public edge) + Host Agents (internal nodes).
    Long-running actions are modeled as Jobs. Real-time notifications via SSE/WebSocket.
servers:
  - url: https://orchestrator.example.com/api/v1
    description: Orchestrator API (public edge)
  - url: http://agent.local:8787/agent/v1
    description: Agent API (internal node)
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /instances/{instanceId}/missions:
    get:
      tags: [Instances, Missions]
      summary: List missions available to instance
      operationId: listMissions
      parameters:
        - $ref: "#/components/parameters/instanceId"
      responses:
        "200":
          description: Missions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Mission" }
...
  /instances/{instanceId}/status:
    get:
      tags: [Agent]
      summary: Local runtime status
      operationId: agentGetInstanceStatus
      servers:
        - url: http://agent.local:8787/agent/v1
      parameters:
        - $ref: "#/components/parameters/instanceId"
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceRuntime"
  /instances/{instanceId}/actions/{action}:
    post:
      tags: [Agent, Jobs]
      summary: Execute a local action (async)
      operationId: agentAction
      servers:
        - url: http://agent.local:8787/agent/v1
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum: [start, stop, restart, update, mission_load, logs_bundle]
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Optional idempotency key to avoid duplicate actions.
...
    instanceId:
      name: instanceId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Problem:
      description: Error response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, example: "https://errors.example.com/instance-not-found" }
        title: { type: string, example: "Instance not found" }
        status: { type: integer, example: 404 }
        detail: { type: string }
        instance: { type: string, description: "Request correlation id / trace id" }
...
    InstanceStatus:
      type: string
      enum: [unknown, stopped, starting, running, stopping, updating, error]
...
    InstanceRuntime:
      type: object
      required: [status, observedAt]
      properties:
        status:
          $ref: "#/components/schemas/InstanceStatus"
        observedAt: { type: string, format: date-time }
        pid: { type: integer, nullable: true }
        startedAt: { type: string, format: date-time, nullable: true }
        uptimeSeconds: { type: integer, nullable: true }
        lastExitCode: { type: integer, nullable: true }
        lastError: { type: string, nullable: true }
...
    Mission:
      type: object
      required: [id, name, path]
      properties:
        id: { type: string, example: "miz_01HXYZ" }
        name: { type: string, example: "Goonfront Caucasus v0.3" }
        path: { type: string, example: "C:\\Missions\\server1\\goonfront_v0_3.miz" }
        sizeBytes: { type: integer, format: int64, nullable: true }
        modifiedAt: { type: string, format: date-time, nullable: true }
...
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, canceled]
...
    JobAccepted:
      type: object
      required: [jobId, status]
      properties:
        jobId: { type: string, example: "job_01HXYZ" }
        status:
          $ref: "#/components/schemas/JobStatus"
...